Created popups for creating and updating usesr profile
