+++ application/config.py
@@ -49,7 +49,7 @@ class Config:
     SKYSMS_PASSWORD_MESSAGE = "Логин:%(login)s\nПароль:%(password)s"
     SKYSMS_MSGCHRSET = 'cyr'
     SKYSMS_MSGENCODING = 'cp1251'
-    SKYSMS_LOGFILE = os.path.join(os.path.abspath(os.path.dirname(os.path.dirname(__file__))), 'log', 'smslog')
+    # SKYSMS_LOGFILE = os.path.join(os.path.abspath(os.path.dirname(os.path.dirname(__file__))), 'log', 'smslog')
 
     session = {
         'session.type': 'file',
