+++ application/templates/admin/layout.html
@@ -12,23 +12,26 @@
     <link rel="stylesheet" href="/static/css/form.css">
     <link rel="stylesheet" href="/static/css/main.css" >
     <link rel="stylesheet" href="/static/css/style.css" >
+    <script src="/static/js/lib/jquery-1.11.3.min.js"></script>
     <script src="/static/jsx/form.jsx" type="text/jsx"></script>
+    <script src="/static/jsx/popup.jsx" type="text/jsx"></script>
     <title>{% block title %}{% endblock %} - PromTal Admin</title>
     {% endblock %}
 </head>
 <body>
+    <div id="popup"></div>
 {% block body %}
     {% include 'admin/header.html' %}
     {% block content %}
     {% endblock %}
 {% endblock %}
-    <script src="/static/js/lib/jquery-1.11.3.min.js"></script>
     <script src="/static/js/lib/jquery.dataTables.min.js"></script>
     <script src="/static/js/lib/dataTables.responsive.min.js"></script>
     <script src="/static/js/lib/dataTables.bootstrap.js"></script>
     <script src="/static/js/lib/bootstrap.js"></script>
     <script src="/static/js/lib/bootstrap-filestyle.min.js"></script>
     <script src="/static/js/lib/react-with-addons.js"></script>
+    <script src="/static/js/lib/rx.all.js"></script>
     <script src="/static/js/lib/JSXTransformer.js"></script>
 </body>
 </html>
